Bit of work on #9, bit of work on #8. Should release this on Monday.
I cleaned up the invocation so that people don't have to know the connection string format anymore.
Also, since the connection string format might not be consistent across drivers, might as well not
make the user depend on that. I need to generate some help message when there is not enough arguments.

I also now take the password securely from the stdin, this is not viable once a service is generated;
I need to implement some type of config-y thing.